pyproject.toml
[tool.poetry]
name = "wb-sch-p2"
version = "0.1.0"
description = ""
authors = ["Vesoore <aleksandrglazkov64962@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.2.3"
numpy = "^2.1.3"
transformers = "^4.46.3"
sentence-transformers = "^3.3.1"
torch = "^2.5.1"
python-dotenv = "^1.0.1"
llama-cpp-python = "^0.3.2"
fastapi = "^0.115.5"
uvicorn = "^0.32.1"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
